Taum and B'day - JS
Taum is planning to celebrate the birthday of his friend, Diksha. There are two types of gifts that Diksha wants from Taum: one is black and the other is white. To make her happy, Taum has to buy b black gifts and w white gifts.

- The cost of each black gift is bc units.

- The cost of every white gift is wc units.

- The cost to convert a black gift into white gift or vice versa is z units.

Determine the minimum cost of Diksha's gifts.